HABA-571 Load memory to m4 SRAM
U-boot 2021 restricts where a file can be loaded into RAM using the
ext4load command. The target location in RAM has to be defined as a
memory in the device-tree.

Load the binary in two steps. Start by loading it to a temporary
address in a valid part of RAM. Copy the binary from the temporary
address to the m4 SRAM mapped region in RAM using the md command.