evco: Use static command status and add update_fdt_eeprom
The board command forwards an i2c error code when an io board is
missing. This is somehow interpreted as a shell exit by uboot, which causes
the fdt selection script to fail.
Always return 1 if the board command fails.

Add the update_fdt_eeprom command, to silence an error at boot.
Having article info in the fdt will not cause any harm.

[EXXT5T8XX-7531]

Signed-off-by: Adrian Marklund <adrian.marklund@teledyneflir.com>