LUKE-683 Add name for cpu1 in imx6q
This is done to be able to use the same dts in u-boot as in Linux.
The name of the cpu is done in u-boot but was not implemented in u-boot.
This commit is a sync to get the linux dts for imx6q.dtsi based dts:es
to work in u-boot.

This commit will probably have to be reverted when u-boot is ported to
a newer version.

Signed-off-by: Jonas Rydow <jonas.rydow@teledyneflir.com>